NUStart Build 1

+++ NUStartApplication/NUStartApplication-ejb/src/java/session/ForumSessionBean.java
@@ -5,11 +5,9 @@
  */
 package session;
 
-import entity.Administrator;
 import entity.Forum;
 import entity.Thread;
 import error.NoResultException;
-import java.util.ArrayList;
 import java.util.List;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
@@ -47,7 +45,6 @@ public class ForumSessionBean implements ForumSessionBeanLocal {
         f.setDescription(f.getDescription().trim());
 
         em.persist(f);
-        f.getCreator().getForums().add(f);
     }
 
     @Override
@@ -61,7 +58,6 @@ public class ForumSessionBean implements ForumSessionBeanLocal {
     @Override
     public void deleteForum(Long fId) throws NoResultException {
         Forum f = getForum(fId);
-        
         em.remove(f);
     }
 
@@ -81,6 +77,7 @@ public class ForumSessionBean implements ForumSessionBeanLocal {
 
     @Override
     public void addThread(Long fId, Thread t) throws NoResultException {
+        t.setTitle(t.getTitle().trim());
         Forum f = getForum(fId);
         em.persist(t);
         f.getThreads().add(t);
@@ -96,8 +93,15 @@ public class ForumSessionBean implements ForumSessionBeanLocal {
     }
 
     @Override
-    public void deleteThread(Long tId) throws NoResultException {
+    public void deleteThread(Long fId, Long tId) throws NoResultException {
+        Forum f = getForum(fId);
         Thread t = threadSessionBeanLocal.getThread(tId);
-        em.remove(t);
+        
+        if (f != null) {
+            f.getThreads().remove(t);
+            em.remove(t);
+        } else {
+            throw new NoResultException("Not found");
+        }
     }
 }
